can header, source split
